+++ ak_library_management/__manifest__.py
@@ -3,7 +3,7 @@
 {
     'name': 'Library Management',
     'author': 'Moin Musa',
-    'version': '18.0.1.2.1',
+    'version': '18.0.1.2.2',
     'summary': 'Manages books,books category and members at library',
     'website': 'https://www.aktivsoftware.com',
     'description': """